% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc-bands.R
\name{adjust_gamma_optimize}
\alias{adjust_gamma_optimize}
\title{Exact gamma for a single sample (L = 1) via dynamic programming.}
\usage{
adjust_gamma_optimize(N, K, conf_level = 0.95)
}
\description{
Exact gamma for a single sample (L = 1) via dynamic programming.
}
\keyword{internal}
