% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basedag.R
\name{basedag_data}
\alias{basedag_data}
\title{Title}
\usage{
basedag_data(
  data_N,
  z1_x_coef,
  z3_x_coef,
  x_y_coef,
  z1_y_coef,
  z2_y_coef,
  y_z4_coef,
  x_z4_coef,
  y_intercept,
  sigma_z1,
  sigma_z2,
  sigma_z3,
  sigma_z4,
  sigma_x,
  sigma_y,
  data_link,
  data_family,
  lb,
  ub,
  resample = 1.3,
  seed = NULL,
  testing_data = TRUE,
  ...
)
}
\arguments{
\item{testing_data}{}
}
\value{

}
\description{
Title
}
