% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basedag.R
\name{basedag_data_noisy}
\alias{basedag_data_noisy}
\title{Title}
\usage{
basedag_data_noisy(
  data_N,
  z1_x_coef,
  z3_x_coef,
  x_y_coef,
  z1_y_coef,
  z2_y_coef,
  y_z4_coef,
  x_z4_coef,
  y_intercept,
  sigma_z1,
  sigma_z2,
  sigma_z3,
  sigma_z4,
  sigma_x,
  sigma_y,
  data_link,
  data_family,
  lb,
  ub,
  resample = 1.3,
  seed = NULL,
  testing_data = TRUE,
  noise_sd = 1,
  noisy_z1_x_coef = NULL,
  noisy_z3_x_coef = NULL,
  noisy_x_y_coef = NULL,
  noisy_z1_y_coef = NULL,
  noisy_z2_y_coef = NULL,
  noisy_y_z4_coef = NULL,
  noisy_x_z4_coef = NULL,
  noisy_y_intercept = NULL,
  noisy_sigma_z1 = NULL,
  noisy_sigma_z2 = NULL,
  noisy_sigma_z3 = NULL,
  noisy_sigma_z4 = NULL,
  noisy_sigma_x = NULL,
  noisy_sigma_y = NULL,
  ...
)
}
\arguments{
\item{noisy_sigma_y}{}
}
\description{
Title
}
