% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesim-package.R
\docType{package}
\name{bayesim-package}
\alias{bayesim}
\alias{bayesim-package}
\title{bayesim: Simulation Framework for Bayesian Modeling}
\description{
bayesim provides a simulation framework for Bayesian modeling studies with
reproducible execution, checkpoint/resume support, and memory-bounded task
processing.
}
\seealso{
\code{\link[=simulation_config]{simulation_config()}}, \code{\link[=run_simulation]{run_simulation()}}, \code{\link[=resume_simulation]{resume_simulation()}},
\link{Fitter}, \link{Metric}
}
\author{
\strong{Maintainer}: Maximilian Scholz \email{research.scholz@mailbox.org}

Authors:
\itemize{
  \item Paul-Christian Bürkner \email{paul.buerkner@gmail.com}
}

}
\keyword{internal}
