% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.R
\name{betaprime}
\alias{betaprime}
\title{Beta prime brms custom family}
\usage{
betaprime(link = "log", link_phi = "log")
}
\arguments{
\item{link}{Link function for function}

\item{link_beta}{Link function for beta argument}
}
\value{
brms beta prime distribution family
}
\description{
Beta prime brms custom family
}
\examples{
# Running the example might take a while and may make RStudio unresponsive.
# Just relax and grab a cup of coffe or tea in the meantime.
a <- rnorm(1000)
data <- list(a = a, y = rbetaprime(1000, exp(0.5 * a + 1), 2))
# BBmisc::surpressAll necassary, the RStudio Roxygen help would be filled with slash symbols...
# For an example without surpress, checkout the Bayesim Betaprime Example script
BBmisc::suppressAll({
  fit1 <- brms::brm(y ~ 1 + a,
    data = data, family = betaprime(),
    stanvars = betaprime()$stanvars, backend = "cmdstanr", cores = 4
  )
})
plot(fit1)
}
