% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{build_simulation_result}
\alias{build_simulation_result}
\title{Build final simulation result}
\usage{
build_simulation_result(config, task_results, task_grid, elapsed)
}
\arguments{
\item{config}{A SimulationConfig S7 object}

\item{task_results}{List of bayesim_task_result objects}

\item{task_grid}{The task grid tibble with updated statuses}

\item{elapsed}{Total elapsed time in seconds}
}
\value{
A bayesim_simulation_result S3 object
}
\description{
Build final simulation result
}
\keyword{internal}
