% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration_result_text}
\alias{calibration_result_text}
\title{Title}
\usage{
calibration_result_text(model_name, results, samples, hypothesis_list)
}
\arguments{
\item{samples}{}
}
\value{

}
\description{
Title
}
