% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration_search}
\alias{calibration_search}
\title{Title}
\usage{
calibration_search(
  x_y_coef_list,
  y_intercept_list,
  link_list,
  likelihood_list,
  sigma_y_list,
  search_list,
  ncores,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
