% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration_text_result}
\alias{calibration_text_result}
\title{Title}
\usage{
calibration_text_result(title, result_df, formula_list)
}
\arguments{
\item{formula_list}{}
}
\value{

}
\description{
Title
}
