% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-config.R
\name{capture_metrics_spec}
\alias{capture_metrics_spec}
\title{Capture Metrics Specification}
\usage{
capture_metrics_spec(metrics)
}
\arguments{
\item{metrics}{A list of Metric objects or NULL.}
}
\value{
A list or NA if NULL.
}
\description{
Creates a hashable representation of metrics configuration.
}
\keyword{internal}
