% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchitnormal.R
\name{cauchitnormal}
\alias{cauchitnormal}
\title{Custom BRMS family Cauchit-Normal in median parametrization.}
\usage{
cauchitnormal(link = "identity", link_sigma = "log")
}
\arguments{
\item{link}{Link function argument (as string) for Median argument. Left as identity!}

\item{link_sigma}{Link function argument (as string) for Shape argument}
}
\value{
Cauchitnormal BRMS model-object
}
\description{
Custom BRMS family Cauchit-Normal in median parametrization.
}
\examples{
library(brms)
library(bayesim)
a <- rnorm(1000)
data <- list(a = a, y = rcauchitnormal(1000, inv_logit_scaled(0.2 + 0.5 * a), 4))
hist(data$y)
fit1 <- brm(y ~ 1 + a, data = data, family = cauchitnormal(),
            stanvars = cauchitnormal()$stanvars, backend = "cmdstan")
plot(fit1)
}
