% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cloglog}
\alias{cloglog}
\title{Complementary-Log-Log link function}
\usage{
cloglog(x)
}
\arguments{
\item{x}{value of x to be transformed, 0 < x < 1}
}
\value{
cloglog value of x
}
\description{
Complementary-Log-Log link function
}
\examples{
x <- seq(from = 0.1 , to = 0.9 , length.out = 100)
y <- cloglog(x)
}
