% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloglognormal.R
\name{cloglognormal_limited}
\alias{cloglognormal_limited}
\title{Custom BRMS family CLogLog-Normal in median parametrization.
Same as Cloglognormal-BRMS, but will limit data to [1e-12, 1 - 1e-12], to improve stability}
\usage{
cloglognormal_limited(link = "identity", link_sigma = "log")
}
\arguments{
\item{link_sigma}{Link function argument (as string) for Shape argument}

\item{linkLink}{function argument (as string) for Median argument. Left as identity!}
}
\value{
Cloglog BRMS model-object
}
\description{
Custom BRMS family CLogLog-Normal in median parametrization.
Same as Cloglognormal-BRMS, but will limit data to [1e-12, 1 - 1e-12], to improve stability
}
\examples{
# Running the example might take a while and may make RStudio unresponsive.
# Just relax and grab a cup of coffe or tea in the meantime.
cloglog_data <- bayesim::rcloglognormal(1000, 0.5, 1)
# cloglognormal_limited will work with values close to the boundaries.
# BBmisc::surpressAll necassary, the RStudio Roxygen help would be filled with slash symbols...
# For an example without surpress, checkout the Bayesim Betaprime Example script
BBmisc::suppressAll({
  fit1 <- brms::brm(y ~ 1,
    data = list(y = cloglog_data), family = bayesim::cloglognormal_limited(),
    stanvars = bayesim::cloglognormal_limited()$stanvars, backend = "cmdstanr", cores = 4
  )
})
plot(fit1)
}
