% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_file_checksum}
\alias{compute_file_checksum}
\title{Compute checksum for a file}
\usage{
compute_file_checksum(path)
}
\arguments{
\item{path}{Character string giving the file path}
}
\value{
Character string containing the MD5 checksum.
}
\description{
Computes an MD5 checksum for a file. Useful for verifying file integrity.
}
\keyword{internal}
