% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_hash}
\alias{compute_hash}
\title{Compute hash of an R object}
\usage{
compute_hash(x)
}
\arguments{
\item{x}{Any R object to hash}
}
\value{
Character string containing the hash value.
}
\description{
Computes a hash of an R object using xxHash64 algorithm for fast,
deterministic hashing.
}
