% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.R
\name{dbetaprime}
\alias{dbetaprime}
\title{Probability density function for the Beta-Prime distribution (aka. inverse Beta)}
\usage{
dbetaprime(x, mu, beta, log = FALSE)
}
\arguments{
\item{x}{Model space, defined for x >= 0, x may be scalar or vector}

\item{mu}{Mean parameter of pdf, mu > 0, mu must be a scalar}

\item{beta}{Second shape parameter of beta-prime, beta > 1 must be a scalar}

\item{log}{Optional argument. If true, returns log(pdf). Normally False.}
}
\value{
pdf of beta-prime, with mean parameterasation
}
\description{
Probability density function for the Beta-Prime distribution (aka. inverse Beta)
}
\examples{
x <- seq(from = 0, to = 100, length.out = 1000)
y <- dbetaprime(x, mu = 4, beta = 2)
plot(x, y, type = "l", ylab = "Density", main = "dbetaprime(mu=4, beta=2)")
}
