% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bank.R
\name{deparse_formula_recursive}
\alias{deparse_formula_recursive}
\title{Recursively deparse a brmsformula to a stable character representation}
\usage{
deparse_formula_recursive(formula)
}
\description{
Recursively deparse a brmsformula to a stable character representation
}
\keyword{internal}
