% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_gamma.R
\name{dgamma_custom}
\alias{dgamma_custom}
\title{Custom Gamma Density function}
\usage{
dgamma_custom(x, mu, a, log = FALSE)
}
\arguments{
\item{x}{Value space, x > 0.}

\item{mu}{Mean parameter of the density, mu > 0.}

\item{k}{Shape parameter, k > 0.}
}
\value{
f(x | mu, k)
}
\description{
Custom Gamma Density function
}
\details{
Define rate constante rho as:
\deqn{\rho(\alpha, \mu) = \frac{\alpha}{\mu}}

The Frechet distribution density is defined as
\deqn{f(y) = \frac{y^{\alpha - 1} exp(-\frac{y}{\rho})} {\rho^\alpha \Gamma(\alpha)} }
}
\examples{
x <- seq(from = 0.01, to = 10, length.out = 1000)
plot(x, dgamma_custom(x, mu = 2, a = 2), type = "l")
}
