% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{dgompertz}
\alias{dgompertz}
\title{Probability density function for the Gompertz distribution, with Median parametrization.}
\usage{
dgompertz(x, mu, beta, log = FALSE)
}
\arguments{
\item{x}{Value}

\item{mu}{Median parameter}

\item{beta}{Scale parameter}

\item{log}{Optional argument. If TRUE, returns log(pdf).}
}
\value{
PDF of gompertz distribution, with median parametrization.
}
\description{
Probability density function for the Gompertz distribution, with Median parametrization.
}
\examples{
x <- seq(from = 0, to = 5, length.out = 100)
y <- bayesim::dgompertz(x, mu = 10, beta = 10)
plot(x, y, type = "l", ylab = "Density", main = "dgompertz(mu=10, beta=10)")
# Compare to online ressources
}
