% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_invgaussian.R
\name{dinversegaussian_custom}
\alias{dinversegaussian_custom}
\title{Probability density function for the Inverse-Gaussion distribution}
\usage{
dinversegaussian_custom(x, mu, shape, log = FALSE)
}
\arguments{
\item{x}{Value space, x > 0.}

\item{mu}{Mean parameter of the density, mu > 0.}

\item{shape}{shape > 0}
}
\value{
f(x | mu, shape)
}
\description{
Probability density function for the Inverse-Gaussion distribution
}
\details{
The PDF is defined with shape parameter alpha as
\deqn{f(y) = \sqrt{\frac{\alpha}{2 \pi y^3}} exp(\frac{-\alpha(y-\mu)^2}{2 \mu^2 y}) }
\url{https://en.wikipedia.org/wiki/Inverse_Gaussian_distribution}
}
\examples{
x <- seq(from = 0.01, to = 10, length.out = 1000)
plot(x, dinversegaussian_custom(x, mu = 2, shape = 2), type = "l")
}
