% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{.fit_spec_response_name}
\alias{.fit_spec_response_name}
\title{Resolve the response name from a fit_spec (LHS of fit_spec$formula).
Returns NA_character_ if it cannot be resolved.}
\usage{
.fit_spec_response_name(fit_spec)
}
\description{
Resolve the response name from a fit_spec (LHS of fit_spec$formula).
Returns NA_character_ if it cannot be resolved.
}
\keyword{internal}
