% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_handler.R
\name{elpd_test}
\alias{elpd_test}
\title{Compute ELPD on test data}
\usage{
elpd_test(fit, newdata, return_object = FALSE)
}
\arguments{
\item{fit}{A brmsfit object}

\item{newdata}{Test data frame}

\item{return_object}{If TRUE, return a custom_loo_object}
}
\value{
A named list with elpd_test and se_elpd_test, or a custom_loo_object
}
\description{
Compute ELPD on test data
}
\examples{
\dontrun{
# Requires brms package
fit <- brms::brm(y ~ x, data = train_data)
elpd_test(fit, test_data)
}
}
