% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{expect_bigger}
\alias{expect_bigger}
\title{Check, if a is bigger than b. expect_smaller(a, b) can be done,
by expect_bigger(b, a) with swapped arguments.}
\usage{
expect_bigger(a, b)
}
\arguments{
\item{a}{number to be tested, should be real scalar or vector}

\item{b}{tested against, should be real scalar or vector}
}
\value{
expect(a > threshold)
}
\description{
For vector/scalar combinations, allowed are:
both scalar
a scalar, b vector -> compare a to all b values
a vector, b scalar -> compare all a values to the threshold b
both vectors -> compare each entry to the other entry
-> different lengths != 1 dissallowed!
}
\examples{
expect_bigger(1, 0) should pass
expect_bigger(0, 1) should fail
expect_bigger(c(1, 2), 0) should pass
expect_bigger(c(0, 2), 1) should fail
expect_bigger(c(1, 2), c(0, 3)) should fail
expect_bigger(c(1, 2, 3), c(0, 1)) should produce an error
}
