% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{expect_eps}
\alias{expect_eps}
\title{Check that |a - b| < eps. Works with scalars and vectors on any input.}
\usage{
expect_eps(a, b, eps)
}
\arguments{
\item{a}{scalar or vector a to be compared}

\item{b}{scalar or vector b to be compared}

\item{eps}{scalar or vector, setting the max differences, eps > 0

For vector/scalar combinations, allowed are:
all scalar
a vector, b scalar, eps scalar -> compare a values to be close enough to b
a scalar, b vector, eps scalar -> compare b values to be close enough to a
a scalar, b scalar, eps vector -> compare the difference a, b too all eps
-> case not intended, but would work
a vector, b scalar, eps vector -> compare each vector-difference entry against each eps
a vector, b scalar, eps vector -> compare each vector-difference entry against each eps
all vectors -> each entry of |a - b| is compared to the same entry in eps
-> different vector lengths != 1 dissallowed!}
}
\value{
expect_true(a close enought to b)
}
\description{
Check that |a - b| < eps. Works with scalars and vectors on any input.
}
\examples{
expect_eps(1, 1.1, 0.2) should pass
expect_eps(-1, -1.1, 0.2) should pass
expect_eps(-1.00001, -1.00001, 1e-4) should pass
expect_eps(1, 1.3, 0.2) should fail
expect_eps(c(1, 1.1), c(1, 1.1, 1.2), 0.5) should produce an error
expect_eps(1, 1.1, -0.2) should produce an error
}
