% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{externalize_artifact}
\alias{externalize_artifact}
\title{Externalize large artifact}
\usage{
externalize_artifact(artifact, artifacts_dir, task_id, field_name)
}
\arguments{
\item{artifact}{An R object to externalize}

\item{artifacts_dir}{Directory path where artifacts should be stored}

\item{task_id}{Task identifier used to construct the filename}

\item{field_name}{Name of the field being externalized (e.g., "draws", "fit")}
}
\value{
A list with pointer information containing:
\itemize{
\item external: TRUE (indicating this is an external pointer)
\item path: Full path to the external file
\item hash: Hash of the artifact for integrity checking
\item size: Size of the artifact in bytes
}
}
\description{
Moves a large artifact to an external RDS file and replaces the in-memory
object with a pointer containing metadata about the externalized file.
}
\seealso{
\code{write_rds_atomic()}, \code{compute_hash()}
}
\keyword{internal}
