% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitter.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to check}

\item{y}{Default value if x is NULL}
}
\value{
x if not NULL, otherwise y
}
\description{
Null coalescing operator
}
