% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs_sbc.R
\name{ifs_SBC}
\alias{ifs_SBC}
\title{Full inverse forward sampling supported SBC}
\usage{
ifs_SBC(
  fit,
  n_sims,
  ppred_data_gen,
  precon_sample = NULL,
  lb = NULL,
  ub = NULL,
  truncate = FALSE,
  post_warmup_samples = 1000,
  ...
)
}
\arguments{
\item{fit}{brmsfit object that is used to generate SBC datasets and fit SBC models}

\item{n_sims}{Number of SBC simulations/datasets.}

\item{ppred_data_gen}{Function that generates the data for the newdata argument of posterior_predict to generate SBC datasets.}

\item{precon_sample}{The dataset that was used to precondition the fit.}

\item{lb}{Lower bound for the outcome. Is used to resample.}

\item{ub}{Upper bound for the outcome. Is used to resample.}

\item{truncate}{If TRUE, will truncate outcome values, if FALSE, will resample instead.}

\item{post_warmup_samples}{Is passed to the sbc model updates and used to
make sure resampling only uses valid samples. Increase in case of numerical instability.}

\item{...}{further parameters. Currently only passed to ppred_data_gen.}
}
\value{
A tibble with SBC summaries and diagnostics.
}
\description{
Full inverse forward sampling supported SBC
}
\examples{
\dontrun{
# Requires SBC and brms packages
}
}
