% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inv_cauchit}
\alias{inv_cauchit}
\title{Inverse Cauchit link function equivalent to pcauchy}
\usage{
inv_cauchit(x)
}
\arguments{
\item{x}{value of x to be transformed, any real scalar or vector allowed}
}
\value{
inverse cauchit of x
}
\description{
Inverse Cauchit link function equivalent to pcauchy
}
\examples{
x <- seq(from = -10, to = 10, length.out = 100)
y <- inv_cauchit(x)
}
