% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inv_cloglog}
\alias{inv_cloglog}
\title{Inverse CLogLog link function}
\usage{
inv_cloglog(x)
}
\arguments{
\item{x}{value of x to be transformed, any real scalar or vector allowed}
}
\value{
inverse-cloglog value of x
}
\description{
Inverse CLogLog link function
}
\examples{
x <- seq(from = -3, to = 1 , length.out = 100)
y <- inv_cloglog(x)
}
