% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse-Logit link function, equal to Logistic link}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{value of x to be transformed, any real scalar or vector allowed}
}
\value{
inverse logit value of x, x e (0, 1)
}
\description{
Inverse-Logit link function, equal to Logistic link
}
\examples{
x <- seq(from = -5, to = 5, length.out = 100)
plot(x, inv_logit(x), type="l")
}
