% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_forward_sampling.R
\name{inverse_forward_sampling}
\alias{inverse_forward_sampling}
\title{Simulate a new dataset using inverse forward sampling.}
\usage{
inverse_forward_sampling(fit, i, n)
}
\arguments{
\item{fit}{An object of class brmsfit.}

\item{i}{The index of a single posterior draw to simulate a dataset for.
The index is passed to posterior_predict's "draw_ids" argument.}

\item{n}{The number of samples for the newly simulated dataset.}
}
\value{
A data.frame containing n observations for each variable in the fit.
}
\description{
Simulate a new dataset using inverse forward sampling.
}
