% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{is_bayesim_error}
\alias{is_bayesim_error}
\title{Check if a condition is a bayesim error}
\usage{
is_bayesim_error(cond)
}
\arguments{
\item{cond}{A condition object to test}
}
\value{
TRUE if the condition is a bayesim error, FALSE otherwise
}
\description{
Check if a condition is a bayesim error
}
\examples{
\dontrun{
tryCatch(
  bayesim_config_error("Invalid config"),
  bayesim_error = function(cond) {
    is_bayesim_error(cond) # TRUE
  }
)
}
}
\keyword{internal}
