% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{is_empty_result_path}
\alias{is_empty_result_path}
\title{Return whether a result path has no user data yet.}
\usage{
is_empty_result_path(result_path)
}
\description{
Return whether a result path has no user data yet.
}
\keyword{internal}
