% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{is_resumable_task_status}
\alias{is_resumable_task_status}
\title{Return whether a task status is eligible for execution on resume.}
\usage{
is_resumable_task_status(status)
}
\description{
Return whether a task status is eligible for execution on resume.
}
\keyword{internal}
