% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-config.R
\name{is_simulation_config}
\alias{is_simulation_config}
\title{Check if Object is a SimulationConfig}
\usage{
is_simulation_config(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
TRUE if x is a SimulationConfig, FALSE otherwise.
}
\description{
Check if Object is a SimulationConfig
}
\keyword{internal}
