% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{isInt_len}
\alias{isInt_len}
\title{Integer vector check}
\usage{
isInt_len(int, len = 1)
}
\arguments{
\item{int}{Integer vector to be checked}

\item{len}{Length of vector, default argument is 1}
}
\value{
Boolean, whether int was Integer and of correct size
}
\description{
Integer vector check
}
\examples{
isInt_len(c(1, 2), 2) # should be TRUE
isInt_len(1) # should be TRUE
isInt_len(1, 2) # should be FALSE, wrong length
isInt_len("r") # should be FALSE, not integer
isInt_len(0.2) # should be FALSE, not integer
isInt_len(c("r", 1), 2) # should be FALSE, partially not integer
isInt_len(c("r", 1), 3) # should be FALSE, both not integer and wrong length
}
