% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{isLogic_len}
\alias{isLogic_len}
\title{Boolean vector check}
\usage{
isLogic_len(logic, len = 1)
}
\arguments{
\item{logic}{Logic vector to be checked}

\item{len}{Length of vector, default argument is 1}
}
\value{
Boolean, whether logic was Boolean and of correct size
}
\description{
Boolean vector check
}
\examples{
isLogic_len(c(TRUE, FALSE), 2) # should be TRUE
isLogic_len(TRUE) # should be TRUE
isLogic_len(TRUE, FALSE) # should be FALSE, wrong length
isLogic_len("r") # should be FALSE, not boolean
isLogic_len(0) # should be FALSE, not boolean
isLogic_len(c("r", TRUE), 2) # should be FALSE, partially not boolean
isLogic_len(c("r", TRUE), 3) # should be FALSE, both not boolean and wrong length
}
