% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{isNat_len}
\alias{isNat_len}
\title{Natural number vector check. Natural in informatic sense (n >= 0).}
\usage{
isNat_len(int, len = 1)
}
\arguments{
\item{int}{Integer vector to be checked}

\item{len}{Length of vector, default argument is 1}
}
\value{
Boolean, whether int was Integer >= 0 and of correct size
}
\description{
Natural number vector check. Natural in informatic sense (n >= 0).
}
\examples{
isNat_len(c(1, 2), 2) # should be TRUE
isNat_len(1) # should be TRUE
isNat_len(-1) # should be FALSE
isNat_len(1, 2) # should be FALSE, wrong length
isNat_len("r") # should be FALSE, not integer
isNat_len(0.2) # should be FALSE, not integer
isNat_len(c("r", 1), 2) # should be FALSE, partially not integer
isNat_len(c("r", 1), 3) # should be FALSE, both not integer and wrong length
}
