% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{isNum_len}
\alias{isNum_len}
\title{Numeric vector check}
\usage{
isNum_len(num, len = 1)
}
\arguments{
\item{num}{Numeric vector to be checked}

\item{len}{Length of vector, default argument is 1}
}
\value{
Boolean, whether num was numeric and of correct size
}
\description{
Numeric vector check
}
\examples{
isNum_len(c(1.1, 2.2), 2) # should be TRUE
isNum_len(0.2) # should be TRUE
isNum_len(0.2, 2) # should be FALSE, wrong length
isNum_len("r") # should be FALSE, not numeric
isNum_len(c("r", 0.2), 2) # should be FALSE, partially not numeric
isNum_len(c("r", 0.2), 3) # should be FALSE, both not numeric and wrong length

}
