% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{isSingleString}
\alias{isSingleString}
\title{Check, if the input is a single string}
\usage{
isSingleString(input)
}
\arguments{
\item{input}{String argument}
}
\value{
Is a string and only one string
}
\description{
Check, if the input is a single string
}
\examples{
isSingleString("abc")  # should be TRUE
isSingleString(c("abc")) # should be TRUE
isSingleString(c("abc", "def")) # should be FALSE, not a single string
isSingleString(1) # should be FALSE, not a string
isSingleString(c("abc", 1)) # should be FALSE, partially not a string and
# also wrong length
}
