% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitter.R
\name{log_lik}
\alias{log_lik}
\title{Compute Pointwise Log-Likelihood}
\usage{
log_lik(fitter, fit_result, newdata = NULL)
}
\arguments{
\item{fitter}{An S7 Fitter object}

\item{fit_result}{A \code{bayesim_fit_result} object from \code{\link[=fit]{fit()}}}

\item{newdata}{Data frame with observations. If NULL, uses training data.}
}
\value{
A matrix with dimensions N x S where:
\itemize{
\item N = number of observations
\item S = number of posterior draws
\item Entry (i, s) is log p(y_i | parameters_s)
}
}
\description{
Compute pointwise log-likelihood values.
}
