% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchitnormal.R
\name{log_lik_cauchitnormal}
\alias{log_lik_cauchitnormal}
\title{Log-Likelihood vignette for the Chauchitnormal distribution, with Median parametrization.}
\usage{
log_lik_cauchitnormal(i, prep)
}
\arguments{
\item{i}{Indices}

\item{prep}{BRMS data}
}
\value{
log_likelihood of the Cauchitnormal distribution, given some BRMS data.
}
\description{
Log-Likelihood vignette for the Chauchitnormal distribution, with Median parametrization.
}
