% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{log_lik_gompertz}
\alias{log_lik_gompertz}
\title{Title}
\usage{
log_lik_gompertz(i, prep)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}
}
\value{
Log-Likelihood of gompertz given data in prep
}
\description{
Title
}
