% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{log_lik_lognormal_custom}
\alias{log_lik_lognormal_custom}
\title{Log-Likelihood vignette for the Lognormal distribution, in Median parametrization.}
\usage{
log_lik_lognormal_custom(i, prep)
}
\arguments{
\item{i}{Indices}

\item{prep}{BRMS data}
}
\value{
Log-Likelihood for BRMS
}
\description{
Log-Likelihood vignette for the Lognormal distribution, in Median parametrization.
}
