% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lomax.R
\name{log_lik_lomax}
\alias{log_lik_lomax}
\title{Title}
\usage{
log_lik_lomax(i, prep)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}
}
\value{
Log-Likelihood of Lomax given data in prep
}
\description{
Title
}
