% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{logistic}
\alias{logistic}
\title{Logistic link function, equivalent to Inverse-Logit link}
\usage{
logistic(x)
}
\arguments{
\item{x}{value of x to be transformed, unbound}
}
\value{
logistic value of x, result ise (0, 1)
}
\description{
Logistic link function, equivalent to Inverse-Logit link
}
\examples{
x <- seq(from = -100, to = 100, length.out = 100)
plot(x, logistic(x), type = "l")
}
