% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitnormal.R
\name{logitnormal}
\alias{logitnormal}
\title{Custom BRMS family Logit-Normal in median parametrization.}
\usage{
logitnormal(link = "identity", link_sigma = "log")
}
\arguments{
\item{link}{Link function argument (as string) for Median argument. Left as identity!}

\item{link_sigma}{Link function argument (as string) for Shape argument}
}
\value{
Logitnormal BRMS model-object
}
\description{
Custom BRMS family Logit-Normal in median parametrization.
}
\examples{
library(brms)
library(bayesim)
a <- rnorm(10000)
data <- list(a = a, y = rlogitnormal(10000, inv_logit_scaled(0.2 + 0.5 * a), 4))
hist(data$y)
fit1 <- brm(y ~ 1 + a, data = data, family = logitnormal(),
            stanvars = logitnormal()$stanvars, backend = "cmdstan")
plot(fit1)
}
