% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{lognormal_custom}
\alias{lognormal_custom}
\title{Custom BRMS family Log-Normal in median parametrization.}
\usage{
lognormal_custom(link = "identity", link_sigma = "log")
}
\arguments{
\item{link}{Link function argument (as string) for Median argument. Left as identity!}

\item{link_sigma}{Link function argument (as string) for Shape argument}
}
\value{
Lognormal BRMS model-object
}
\description{
Custom BRMS family Log-Normal in median parametrization.
}
\examples{
library(brms)
a <- rnorm(1000)
data <- list(a = a, y = rlognormal_custom(n, exp(0.5 * a + 1), 2))
fit1 <- brm(y ~ 1 + a, data = data, family = lognormal_custom(),
  stanvars = lognormal_custom()$stanvars, backend = "cmdstan")
plot(fit1)
}
