% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{new_run_policy}
\alias{new_run_policy}
\title{Construct the operational policy for a simulation run.}
\usage{
new_run_policy(config, progress = TRUE, workers = NULL, verbose = TRUE)
}
\description{
Construct the operational policy for a simulation run.
}
\keyword{internal}
