% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{new_study_spec}
\alias{new_study_spec}
\title{Construct the immutable scientific part of a simulation configuration.}
\usage{
new_study_spec(config)
}
\description{
Construct the immutable scientific part of a simulation configuration.
}
\keyword{internal}
