% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_forward_sampling.R
\name{nodes_by_depth}
\alias{nodes_by_depth}
\title{Determine depth of nodes in dependency graph}
\usage{
nodes_by_depth(adj_matrix)
}
\arguments{
\item{adj_matrix}{Adjacency matrix representing variable dependencies}
}
\value{
List of nodes grouped by their depth in the dependency graph
}
\description{
Determine depth of nodes in dependency graph
}
\keyword{internal}
