% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bank.R
\name{normalize_family}
\alias{normalize_family}
\title{Normalize a brms family to a hashable list}
\usage{
normalize_family(family)
}
\description{
Normalize a brms family to a hashable list
}
\keyword{internal}
