% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bank.R
\name{normalize_prior}
\alias{normalize_prior}
\title{Normalize a brms prior object to a hashable data.frame representation}
\usage{
normalize_prior(prior)
}
\description{
Normalize a brms prior object to a hashable data.frame representation
}
\keyword{internal}
