% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bank.R
\name{normalize_stanvars}
\alias{normalize_stanvars}
\title{Normalize a brms stanvars object to a hashable representation}
\usage{
normalize_stanvars(stanvars)
}
\description{
Normalize a brms stanvars object to a hashable representation
}
\keyword{internal}
