% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc-bands.R
\name{p_interior}
\alias{p_interior}
\title{Interior-probability recursion helper.}
\usage{
p_interior(p_int, x1, x2, z1, z2, gamma, N)
}
\description{
Interior-probability recursion helper.
}
\keyword{internal}
