% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{Title}
\usage{
plot_dist(
  dist,
  bounds,
  pars,
  prefix = "d",
  parnames = NULL,
  package = NULL,
  user_theme = ggthemes::theme_tufte,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
