% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_beta.R
\name{posterior_epred_beta}
\alias{posterior_epred_beta}
\title{Posterior expected value prediction of the custom-beta implementation.}
\usage{
posterior_epred_beta(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Recover the given mean of data prep
}
\description{
Posterior expected value prediction of the custom-beta implementation.
}
