% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.R
\name{posterior_epred_betaprime}
\alias{posterior_epred_betaprime}
\title{Posterior-Expectation-Predict vignette for the Beta-Prime distribution, in mean parametrization.}
\usage{
posterior_epred_betaprime(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Recover the given mean of data prep
}
\description{
Posterior-Expectation-Predict vignette for the Beta-Prime distribution, in mean parametrization.
}
