% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_invgaussian.R
\name{posterior_epred_inversegaussian_custom}
\alias{posterior_epred_inversegaussian_custom}
\title{Posterior-Expectation-Predict vignette for the Inverse-Gaussian distribution, in mean parametrization.}
\usage{
posterior_epred_inversegaussian_custom(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Recover the given mean of data prep
}
\description{
Posterior-Expectation-Predict vignette for the Inverse-Gaussian distribution, in mean parametrization.
}
