% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{posterior_epred_lognormal}
\alias{posterior_epred_lognormal}
\title{Posterior expected value prediction vignette for Lognormal distribution.}
\usage{
posterior_epred_lognormal(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Mean of Posterior
}
\description{
Posterior expected value prediction vignette for Lognormal distribution.
}
