% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{posterior_epred_lognormal_custom}
\alias{posterior_epred_lognormal_custom}
\title{Posterior expected value prediction vignette for Lognormal distribution.}
\usage{
posterior_epred_lognormal_custom(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Expectation value prediction for BRMS
}
\description{
Posterior expected value prediction vignette for Lognormal distribution.
}
