% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lomax.R
\name{posterior_epred_lomax}
\alias{posterior_epred_lomax}
\title{Title}
\usage{
posterior_epred_lomax(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Recover the given mean of data prep
}
\description{
Title
}
