% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz_comparisons.R
\name{posterior_predict_gompertz_ab}
\alias{posterior_predict_gompertz_ab}
\title{Title}
\usage{
posterior_predict_gompertz_ab(i, prep, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
