% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz_comparisons.R
\name{posterior_predict_gompertz_b}
\alias{posterior_predict_gompertz_b}
\title{Title}
\usage{
posterior_predict_gompertz_b(i, prep, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
