% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_beta.R
\name{qbeta_custom}
\alias{qbeta_custom}
\title{Mean parameterization of the beta quantile function}
\usage{
qbeta_custom(p, mu, phi)
}
\arguments{
\item{p}{Probabilities, for which to calculate the quantiles}

\item{mu}{Mean, mu e (0, 1).}

\item{phi}{Precision parameter, phi > 0.}
}
\value{
Quantiles of the beta distribution
}
\description{
Mean parameterization of the beta quantile function
}
\examples{
x <- seq(from = 0, to = 1, length.out = 100)
plot(x, qbetaprime(x, mu = 1, phi = 2), type = "l")
}
