% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.R
\name{qbetaprime}
\alias{qbetaprime}
\title{Title}
\usage{
qbetaprime(p, mu, beta)
}
\arguments{
\item{p}{Probabilities, for which to calculate the quantiles}

\item{mu}{Mean of the beta-prime, mu > 0}

\item{beta}{Beta Shape Argument of beta-prime, beta > 1}
}
\value{
Quantiles of the beta-prime distribution, given p, mu and beta
}
\description{
Title
}
\examples{
x <- seq(from = 0, to = 1, length.out = 100)
}
