% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_handler.R
\name{r2}
\alias{r2}
\title{Calculate pointwise R-squared}
\usage{
r2(y, yrep, weights = NULL)
}
\arguments{
\item{y}{Vector of observed values}

\item{yrep}{Matrix of predicted values (draws x observations)}

\item{weights}{PSIS weights (optional)}
}
\value{
Pointwise R-squared values
}
\description{
Calculate pointwise R-squared
}
\examples{
\dontrun{
y <- c(1, 2, 3)
yrep <- matrix(rnorm(300), nrow = 100, ncol = 3)
r2(y, yrep)
}
}
\keyword{internal}
