% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_handler.R
\name{r2_loo}
\alias{r2_loo}
\title{Calculate PSIS-loo R² for a given brms fit}
\usage{
r2_loo(fit, psis_object = NULL, yrep = NULL, return_object = FALSE, ...)
}
\arguments{
\item{fit}{brms fit to calculate rmse for}

\item{...}{}
}
\value{
\code{custom_loo_object} object with R² acting as elpd.
}
\description{
Calculate PSIS-loo R² for a given brms fit
}
\examples{
\dontrun{
}
}
