% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta.R
\name{rBeta}
\alias{rBeta}
\title{RNG for the beta distribution using brms' mean and phi parametrization.}
\usage{
rBeta(n, mu, phi)
}
\arguments{
\item{phi}{}
}
\value{

}
\description{
RNG for the beta distribution using brms' mean and phi parametrization.
}
