% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.R
\name{rbetaprime}
\alias{rbetaprime}
\title{Title}
\usage{
rbetaprime(n, mu, beta)
}
\arguments{
\item{n}{Number of beta-prime samples, to be calculated for RNG.}

\item{mu}{Mean of beta-prime distribution}

\item{beta}{Beta shape argument, of beta-prime}
}
\value{
Randomly samples distributed as beta-prime-
}
\description{
Title
}
\examples{
y <- bayesim::rbetaprime(100, mu = 1, beta = 2)
}
