% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint.R
\name{read_outcome_shards}
\alias{read_outcome_shards}
\title{Validate and optionally materialize accumulated canonical outcome shards.}
\usage{
read_outcome_shards(result_path, refs, load_outcomes = TRUE)
}
\description{
Validate and optionally materialize accumulated canonical outcome shards.
}
\keyword{internal}
