% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint.R
\name{read_redundant_shard}
\alias{read_redundant_shard}
\title{Read a checksummed shard, falling back to its mirror.}
\usage{
read_redundant_shard(directory, ref)
}
\description{
Read a checksummed shard, falling back to its mirror.
}
\keyword{internal}
