% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\docType{data}
\name{RETENTION_CONTEXTS}
\alias{RETENTION_CONTEXTS}
\title{Valid retention contexts}
\format{
Character vector
}
\usage{
RETENTION_CONTEXTS
}
\description{
Valid retention contexts
}
\keyword{internal}
