% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_gamma.R
\name{rgamma_custom}
\alias{rgamma_custom}
\title{Custom Gamma RNG function in mean parametrisation.}
\usage{
rgamma_custom(n, mu, a)
}
\arguments{
\item{n}{Number of samples, scalar natural number.}

\item{mu}{Mean parameter, mu > 0.}

\item{k}{Shape parameter, k > 0.}
}
\value{
n Gamma distributed samples.
}
\description{
Custom Gamma RNG function in mean parametrisation.
}
\examples{
hist(log(rgamma_custom(10000, mu = 2, a = 1)))
}
