% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{rlognormal_custom}
\alias{rlognormal_custom}
\title{Lognormal RNG-function in median parametrization.}
\usage{
rlognormal_custom(n, mu, sigma)
}
\arguments{
\item{n}{Number of draws}

\item{mu}{Median paramameter, mu unbound, mu already log transformed}

\item{sigma}{Sigma shape parameter, sigma > 0}
}
\value{
n Lognormally ditributed samples
}
\description{
Lognormal RNG-function in median parametrization.
}
