% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_handler.R
\name{rmse_loo}
\alias{rmse_loo}
\title{Calculate PSIS-loo rmse for a given brms fit}
\usage{
rmse_loo(fit, psis_object = NULL, return_object = FALSE, yrep = NULL, ...)
}
\arguments{
\item{psis_object}{}

\item{...}{Additional arguments to be passed to update() in case of reloo}
}
\value{
\code{custom_loo_object} object with rmse acting as elpd.
}
\description{
Calculate PSIS-loo rmse for a given brms fit
}
\examples{
\dontrun{
}
}
