% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_weibull.R
\name{rweibull_custom}
\alias{rweibull_custom}
\title{Custom Weibull RNG function in mean parametrisation.}
\usage{
rweibull_custom(n, mu, k)
}
\arguments{
\item{n}{Number of samples, scalar natural number.}

\item{mu}{Mean parameter, mu > 0.}

\item{k}{Shape parameter, k > 0.}
}
\value{
n Weibull distributed samples.
}
\description{
Custom Weibull RNG function in mean parametrisation.
}
\examples{
hist(log(rweibull_custom(10000, mu=2, k=1)))
}
