% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bank.R
\name{stan_threads_arg}
\alias{stan_threads_arg}
\title{Extract the \code{threads} argument from stan_args, if present}
\usage{
stan_threads_arg(stan_args)
}
\description{
Extract the \code{threads} argument from stan_args, if present
}
\keyword{internal}
