% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{test_brms_quantile}
\alias{test_brms_quantile}
\title{Check, that data of the posterior is close enough to the reference data.}
\usage{
test_brms_quantile(posterior_data, name, reference, thresh, debug = FALSE)
}
\arguments{
\item{posterior_data}{Data fitted and drawn}

\item{name}{Name of the variable to check}

\item{reference}{Reference value to check against}

\item{thresh}{Scalar or 2-length vector of quantile bounds.
For scalar constructs bound as [tresh, 1-thresh]}
}
\value{
succeess, fail or error
}
\description{
Check, that data of the posterior is close enough to the reference data.
}
\examples{
n <- 1000
a <- rnorm(n)
ba_in <- 0.5
data <- list(a = a, y = bayesim::rbetaprime(n, exp(ba_in * a + 1), 2))
fit1 <- brms::brm(y ~ 1 + a,
  data = data, family = bayesim::betaprime(),
  stanvars = bayesim::betaprime()$stanvars, backend = "cmdstan",
  cores = 4, silent = 2, refresh = 0
)
test_brms_quantile(fit1, "b_a", ba_in, 0.025)
# TODO: how to test this function (and is this to be tested)?
}
