% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{test_rng}
\alias{test_rng}
\title{Distribution RNG test vingette}
\usage{
test_rng(
  rng_fun,
  metric_mu,
  n,
  mus,
  shapes,
  mu_eps,
  p_acceptable_failures,
  mu_link = identity
)
}
\arguments{
\item{rng_fun}{RNG function under test}

\item{metric_mu}{Metric to be used on RNG data (usually mean or median)}

\item{mus}{Metric data used as RNG argument and to be compared to (usually mean or median)}

\item{shapes}{Shape arguemnts}

\item{mu_eps}{Acceptable difference of |mu - metric_mu(rng_fun)}

\item{p_acceptable_failures}{Acceptable rate of failure, relative value of difference bigger mu_eps}

\item{mu_link}{Default=identity, optional link-function argument, for example
useful in link-normal-distributions}
}
\value{
Nothing actually, just wraps the test
}
\description{
Distribution RNG test vingette
}
\examples{
library(testthat)
library(bayesim)
mus <- seq(from = 1 + eps, to = 20, length.out = 10)
phis <- seq(from = 2 + eps, to = 20, length.out = 10)
test_rng(
  rng_fun = bayesim::rbetaprime, metric_mu = mean, n = 10000, mus = mus,
  shapes = phis, mu_eps = 0.2, p_acceptable_failures = 0.05
)
}
