% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{vector_combinator}
\alias{vector_combinator}
\title{Vector combinator. Used in a few link-normal tests (to simplify indexing, hehe).
Not meant to be used by users. (Hence no at-export)}
\usage{
vector_combinator(a, b)
}
\arguments{
\item{a}{Vector a}

\item{b}{Vector b}
}
\value{
combination c(c(a[1], b[1]), c(a[1], b[2]), ...)
}
\description{
Vector combinator. Used in a few link-normal tests (to simplify indexing, hehe).
Not meant to be used by users. (Hence no at-export)
}
\examples{
print(bayesim:::vector_combinator(seq(0, 1, length.out=3), seq(0, 2, length.out=3)))
}
