% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basedag.R
\name{basedag_data}
\alias{basedag_data}
\title{Title}
\usage{
basedag_data(
  data_N,
  z1_x_coef = 1,
  z3_x_coef = 1,
  x_y_coef = 1,
  z1_y_coef = 1,
  z2_y_coef = 1,
  y_z4_coef = 1,
  x_z4_coef = 1,
  y_intercept = 0,
  sigma_z1 = 1,
  sigma_z2 = 1,
  sigma_z3 = 1,
  sigma_z4 = 1,
  sigma_x = 1,
  sigma_y = 1,
  data_link,
  RNG,
  lb,
  ub,
  oversampling = 10
)
}
\arguments{
\item{oversampling}{}
}
\value{

}
\description{
Title
}
