% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-built-in.R
\name{BiasMetric}
\alias{BiasMetric}
\alias{pred_bias_metric}
\title{Bias Metric}
\usage{
BiasMetric(name = character(0), needs = character(0), required = FALSE)

pred_bias_metric(name = "bias")
}
\arguments{
\item{name}{Character string naming the metric. Defaults to "bias".}

\item{needs}{Character vector of required capabilities. Defaults to "predictions".}

\item{required}{Logical indicating if metric failure causes task failure. Defaults to FALSE.}
}
\value{
A BiasMetric object.

A BiasMetric object.
}
\description{
Mean bias of predictions.

Constructor for BiasMetric.
}
\keyword{internal}
