% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_forward_sampling.R
\name{brms_full_ppred}
\alias{brms_full_ppred}
\title{Full forward sampling of a the response of brms fit, including multivariate models.}
\source{
This function is taken from the SBC package
  (https://github.com/hyunjimoon/SBC) and only here to ensure stability, as
  it is not exported in SBC.
}
\usage{
brms_full_ppred(fit, newdata = NULL, draws = NULL, validate_all = FALSE)
}
\arguments{
\item{fit}{An object of class `brmsfit`}

\item{newdata}{An optional data.frame for which to evaluate predictions. If NULL (default), the original data of the model is used.}

\item{draws}{An integer vector specifying the posterior draws to be used. If NULL (the default), all draws are used.}

\item{validate_all}{if TRUE, validation of input data will be done in all iterations, otherwise only once}
}
\value{
A list of data.frames containing the draws.
}
\description{
Full forward sampling of a the response of brms fit, including multivariate models.
}
\examples{
# Pending
}
\keyword{internal}
