% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_forward_sampling.R
\name{brms_response_sequence}
\alias{brms_response_sequence}
\title{Determine the response sequence of brms model}
\source{
This function is taken from the SBC package
(https://github.com/hyunjimoon/SBC) and only here to ensure stability, as
it is not exported in SBC.
}
\usage{
brms_response_sequence(x)
}
\description{
Determine the response sequence of brms model
}
\keyword{internal}
