% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms-fitter.R
\name{BrmsFitter}
\alias{BrmsFitter}
\title{Brms Fitter}
\usage{
BrmsFitter(
  name = character(0),
  supports_predictions = TRUE,
  supports_log_lik = TRUE,
  supports_loo = TRUE,
  backend = "rstan",
  chains = 4L,
  iter = 2000L,
  warmup = 1000L,
  thin = 1L,
  refresh = 0L,
  silent = 2L,
  cores = 1L
)
}
\description{
Fitter implementation for brms models.
}
