% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefit.R
\name{build_prefit_list}
\alias{build_prefit_list}
\title{Prepare a list of brmsfit objects to update during repeated simulations}
\usage{
build_prefit_list(fit_configuration, stan_pars)
}
\arguments{
\item{fit_configuration}{A named list that currently holds family, link and
prior.}

\item{stan_pars}{A named list which contains a backend field.}
}
\value{
A named list of precompiled fit objects keyed by fit configuration.
}
\description{
This is mainly used to save on constant recompilation times.
}
\keyword{internal}
