% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helpers.R
\name{cluster_setup}
\alias{cluster_setup}
\title{Convenience Function to set up a cluster used for multiprocessing}
\usage{
cluster_setup(ncores = 2, cluster_type = "FORK", debug = FALSE, outfile = NULL)
}
\arguments{
\item{ncores}{Numbers of processes the cluster should use}

\item{cluster_type}{"FORK" or "PSOCK". Fork is faster but doesn't work on
Windows}

\item{debug}{TRUE if the cluster log should be written to file}

\item{outfile}{Path where the cluster log is written to in debug mode.}
}
\description{
Convenience Function to set up a cluster used for multiprocessing
}
