% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{construct_brms}
\alias{construct_brms}
\title{Title}
\usage{
construct_brms(
  n_data_sampels,
  ba,
  intercept,
  shape,
  link,
  family,
  rng,
  num_parallel,
  seed = NULL,
  suppress_output = TRUE,
  data_threshold = NULL
)
}
\arguments{
\item{data_threshold}{}
}
\description{
Title
}
