% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-built-in.R
\name{CoverageMetric}
\alias{CoverageMetric}
\title{Coverage Metric}
\usage{
CoverageMetric(
  name = character(0),
  needs = character(0),
  required = FALSE,
  prob = 0.95
)
}
\description{
Coverage of true parameter values within credible intervals.
}
