% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{dataset_conf_sim}
\alias{dataset_conf_sim}
\title{Title}
\usage{
dataset_conf_sim(
  data_gen_conf,
  fit_confs,
  numeric_metrics,
  predictive_metrics,
  prefits,
  seed = NULL,
  path = NULL,
  brms_backend,
  cmdstanr_path,
  ncores
)
}
\arguments{
\item{ncores}{}
}
\value{

}
\description{
Title
}
