% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{dataset_conf_sim}
\alias{dataset_conf_sim}
\title{Title}
\usage{
dataset_conf_sim(
  data_gen_conf,
  data_gen_fun,
  fit_confs,
  prefits,
  seed = NULL,
  result_path = NULL,
  stan_pars,
  ncores,
  cluster_type,
  debug,
  global_seed,
  ...
)
}
\arguments{
\item{debug}{}
}
\description{
Title
}
