% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{dataset_sim}
\alias{dataset_sim}
\title{Title}
\usage{
dataset_sim(
  data_gen_conf,
  fit_confs,
  prefits,
  numeric_metrics,
  predictive_metrics,
  brms_backend,
  cmdstan_path,
  cmdstan_write_path,
  seed,
  debug,
  result_path
)
}
\arguments{
\item{path}{}
}
\description{
Title
}
