% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{dataset_sim}
\alias{dataset_sim}
\title{Title}
\usage{
dataset_sim(
  data_gen_conf,
  data_gen_fun,
  fit_confs,
  prefits,
  stan_pars,
  seed,
  debug,
  result_path,
  ...
)
}
\arguments{
\item{path}{}
}
\description{
Title
}
