% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{density_lookup_generator}
\alias{density_lookup_generator}
\title{Yannick's super dooper density lookup data generator!
This is meant for development only. Use at your own risk (and sanity). You have been warned!}
\usage{
density_lookup_generator(
  n_param = 10,
  n_x = 100,
  eps = 10^-6,
  mu_int,
  shape_int,
  x_int_prelink,
  x_link = identity,
  density_fun,
  density_name
)
}
\arguments{
\item{n_param}{Number of paramerter for each mu and shape. Will yield n_param^2 permutations of both.}

\item{n_x}{Number of the x values.}

\item{eps}{Allowed precision tolerance.}

\item{mu_int}{Interval of the densities mu variable.}

\item{shape_int}{Interval of the densities shape variable.}

\item{x_int_prelink}{Interval of x-values, before the x_link function is used on them.}

\item{x_link}{Link function to be used on x_int_prelink. Default is identity.}

\item{density_fun}{The PDF to be saved.}

\item{density_name}{The name of the whole pre-calculated thingy. Probably the one of the density.}
}
\value{
Nothing, but saves input and reference files for later use!
}
\description{
Yannick's super dooper density lookup data generator!
This is meant for development only. Use at your own risk (and sanity). You have been warned!
}
\examples{
library(bayesim)
eps <- 1e-6
density_lookup_generator  (mu_int=c(eps, 1-eps), shape_int=c(2,10), x_int_prelink=c(eps, 1-eps),
                          density_fun=bayesim::dcauchitnormal, density_name="cauchitnormal_demodata")
}
