% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{dgompertz}
\alias{dgompertz}
\title{Probability density function for the Gompertz distribution, with Median parametrization.}
\usage{
dgompertz(x, mu, eta, log = FALSE)
}
\arguments{
\item{x}{Model space, defined for x >= 0}

\item{mu}{Median parameter of pdf, mu > 0}

\item{eta}{Second shape parameter of Gompertz, defined for eta > 1}

\item{log}{Optional argument. If TRUE, returns log(pdf). Normally False.}
}
\value{
f(x | mu, eta)
}
\description{
Probability density function for the Gompertz distribution, with Median parametrization.
}
\details{
PDF of Gompertz implementation, with constant b:
\deqn{b(\mu,\eta) := (1 / \mu) * log1p((-1 / \eta) * log(0.5))}
\deqn{f(x) = \eta*b*exp(\eta + bx - \eta * e^{bx})}
}
\examples{
x <- seq(from = 0, to = 10, length.out = 100)
plot(x, dgompertz(x, mu=2, eta=0.2), type = "l")
}
