% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{dgompertz}
\alias{dgompertz}
\title{Title}
\usage{
dgompertz(x, mu, eta, log = FALSE)
}
\arguments{
\item{x}{Model space, defined for x >= 0}

\item{mu}{Median parameter of pdf, mu > 0}

\item{eta}{Second shape parameter of Gompertz, defined for eta > 1}

\item{log}{Optional argument. If true, returns log(pdf). Normally False.}
}
\value{
pdf of gompertz distribution, with median parameterasation
}
\description{
Title
}
\examples{
x <- seq(from = 0, to = 5, length.out = 100)
eta <- 0.1
b <- 1
median <- (1 / b) * log((-1 / eta) * log(1 / 2) + 1)
y <- bayesim::dgompertz(x, mu = median, eta = eta)
plot(x, y, type = "l", ylab = "Density", main = "dgompertz(mu=2.0708, eta=0.1) or dgompertz(b=1, eta=0.1)")
# Compare to online ressources
}
