% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_handler.R
\name{elpd_pointwise_summaries}
\alias{elpd_pointwise_summaries}
\title{Pointwise elps summaries}
\usage{
elpd_pointwise_summaries(fit, quantiles, newdata = NULL)
}
\arguments{
\item{fit}{A brmsfit object.}

\item{quantiles}{A vector of quantiles of interest.}

\item{newdata}{If supplied, returns the summaries for \code{\link[=elpd_test]{elpd_test()}}
otherwise, returns \code{elpd} summaries by
default.}
}
\value{
A named list of summaries.
}
\description{
Convenience function to collect quantiles and summaries of the pointwise elpd
estimates instead of just the main estimates.
The returned summaries are all sample size independent.
}
\examples{
\dontrun{
fit <- brms::brm(y ~ 1, data = list(rnorm(1000)))
elpd_pointwise_summaries(fit, seq(0.1, 0.9, length.out = 9))
elpd_pointwise_summaries(fit, seq(0.1, 0.9, length.out = 9), rnorm(1000))
}
}
