% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{estimate_size}
\alias{estimate_size}
\title{Estimate object size}
\usage{
estimate_size(x)
}
\arguments{
\item{x}{An R object to estimate}
}
\value{
Size in bytes as numeric
}
\description{
Estimates the memory size of an R object in bytes.
}
\examples{
estimate_size(1:1000)
estimate_size(iris)
}
