% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{expect_brms_family}
\alias{expect_brms_family}
\title{Title}
\usage{
expect_brms_family(
  n_data_sampels = 1000,
  ba = 0.5,
  intercept = 1,
  shape = 2,
  link,
  family,
  rng,
  shape_name,
  thresh = 0.05,
  num_parallel = 2,
  debug = FALSE,
  data_threshold = NULL,
  seed = 1337
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
