% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms-fitter.R
\name{extract_brms_diagnostics}
\alias{extract_brms_diagnostics}
\title{Extract brms diagnostics}
\usage{
extract_brms_diagnostics(fit)
}
\arguments{
\item{fit}{brms fit object}
}
\value{
Named list of diagnostics
}
\description{
Extract brms diagnostics
}
