% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_lookup.R
\name{fit_conf_key}
\alias{fit_conf_key}
\title{Generate lookup keys for fit configurations to retrieve prefit objects
matching said config.}
\usage{
fit_conf_key(fit_conf)
}
\arguments{
\item{fit_conf}{}
}
\value{
A hash generated from the fit configuration
}
\description{
Generate lookup keys for fit configurations to retrieve prefit objects
matching said config.
}
\examples{
fit_conf_key(
  list(
    fit_family = "gaussian",
    fit_link = "identity",
    prior = list(c(brms::set_prior("", class = "Intercept")))
  )
)
}
