% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{fit_sim}
\alias{fit_sim}
\title{Title}
\usage{
fit_sim(
  prefit,
  dataset,
  data_gen_conf,
  fit_conf,
  seed,
  debug,
  result_path,
  stan_pars,
  ...
)
}
\arguments{
\item{path}{}
}
\description{
Title
}
