% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{fit_sim}
\alias{fit_sim}
\title{Title}
\usage{
fit_sim(
  prefit,
  dataset,
  testing_data,
  numeric_metrics,
  predictive_metrics,
  data_gen_conf,
  fit_conf,
  brms_backend,
  seed,
  debug,
  path
)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Title
}
