% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{full_simulation}
\alias{full_simulation}
\title{Title}
\usage{
full_simulation(
  data_gen_confs,
  fit_confs,
  numeric_metrics,
  predictive_metrics,
  ncores_prefit = 1,
  ncores_simulation = 1,
  brms_backend = "cmdstanr",
  seed = NULL,
  path = NULL
)
}
\arguments{
\item{fit_confs}{}
}
\value{

}
\description{
Title
}
