% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{full_simulation}
\alias{full_simulation}
\title{Title}
\usage{
full_simulation(
  data_gen_confs,
  fit_confs,
  numeric_metrics,
  predictive_metrics,
  ncores_simulation = 1,
  brms_backend = "cmdstanr",
  cmdstan_path = NULL,
  cmdstan_write_path = NULL,
  seed = NULL,
  result_path = NULL,
  debug = FALSE
)
}
\arguments{
\item{debug}{}
}
\description{
Title
}
