% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{full_simulation}
\alias{full_simulation}
\title{Title}
\usage{
full_simulation(
  data_gen_confs,
  fit_confs,
  metrics,
  ncores,
  seed = NULL,
  path = NULL
)
}
\arguments{
\item{metrics}{}
}
\value{

}
\description{
Title
}
