% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{full_simulation}
\alias{full_simulation}
\title{Title}
\usage{
full_simulation(
  data_gen_confs,
  fit_confs,
  ncores_simulation = 1,
  cluster_type = "PSOCK",
  stan_pars,
  seed = NULL,
  result_path = NULL,
  debug = FALSE,
  ...
)
}
\arguments{
\item{debug}{}
}
\description{
Title
}
