% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs_sbc.R
\name{ifs_SBC}
\alias{ifs_SBC}
\title{Full inverse forward sampling supported SBC}
\usage{
ifs_SBC(
  fit,
  n_sims,
  ppred_data_gen,
  precon_sample = NULL,
  lb = NULL,
  ub = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Full inverse forward sampling supported SBC
}
