% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_invgaussian.R
\name{inversegaussian_custom}
\alias{inversegaussian_custom}
\title{Custom Inverse-Gaussian BRMS-implementation in mean parametrization.}
\usage{
inversegaussian_custom(link = "log", link_shape = "log")
}
\arguments{
\item{link}{Link function for function}

\item{link_shape}{Link function for the shape argument}
}
\value{
BRMS inversegaussian distribution family
}
\description{
Custom Inverse-Gaussian BRMS-implementation in mean parametrization.
}
\examples{
# Running the example might take a while and may make RStudio unresponsive.
# Just relax and grab a cup of coffe or tea in the meantime.
a <- rnorm(1000)
data <- list(a = a, y = bayesim::rinversegaussian_custom(1000, exp(0.5 * a + 1), 2))
# BBmisc::surpressAll necassary, the RStudio Roxygen help would be filled with slash symbols...
# For an example without surpress, checkout the Bayesim Betaprime Example script
BBmisc::suppressAll({  fit1 <- brms::brm(y ~ 1 + a, data = data, family = bayesim::inversegaussian_custom(),
  stanvars = bayesim::inversegaussian_custom()$stanvars, backend = "cmdstanr", cores = 4)  })
plot(fit1)
}
