% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{is_bayesim_simulation_result}
\alias{is_bayesim_simulation_result}
\title{Check if object is a bayesim_simulation_result}
\usage{
is_bayesim_simulation_result(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"bayesim_simulation_result"}, \code{FALSE} otherwise
}
\description{
Check if object is a bayesim_simulation_result
}
\keyword{internal}
