% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaprime.R
\name{log_lik_betaprime}
\alias{log_lik_betaprime}
\title{Log-Likelihood of the beta prime distribution}
\usage{
log_lik_betaprime(i, prep)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}
}
\value{
Log-Likelihood of betaprime given data in prep
}
\description{
Log-Likelihood of the beta prime distribution
}
