% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_handler.R
\name{loo_compare_handler}
\alias{loo_compare_handler}
\title{Title}
\usage{
loo_compare_handler(loo_object_matrix, predictive_metrics)
}
\arguments{
\item{predictive_metrics}{}
}
\description{
Title
}
\examples{
\dontrun{
}
}
