% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_list_handler.R
\name{metric_list_handler}
\alias{metric_list_handler}
\title{Collect multiple metrics at once}
\usage{
metric_list_handler(fit, metrics, data_gen_output, fit_conf, ...)
}
\arguments{
\item{fit}{A brmsfit object.}

\item{metrics}{A vector of metric identifiers. See details for supported
identifiers.}

\item{data_gen_output}{Output from data generation containing true parameters}

\item{fit_conf}{Model configuration list or data.frame row}

\item{...}{Additional arguments passed to metric calculation functions}
}
\value{
A named list containing all requested metrics' results.
}
\description{
A convenience function that collects all given metrics at once. This
can save time compared to manually calling all metric functions
individually, as some variables can be reused instead of being calculated
multiple times.
}
\details{
Currently, the following identifiers are supported:
\itemize{
\item "bias": posterior bias
\item "divergents": number of divergent transitions
\item "ess": effective sample size
\item "elpd_loo": ELPD-LOO
\item "elpd_newdata": ELPD on new data
\item "epred": expected predictions
\item "mae_s": mean absolute error scaled
\item "p_mean": posterior mean
\item "p_sd": posterior standard deviation
\item "pareto_k": Pareto k diagnostic values
\item "pos_prob": posterior probability of positive values
\item "ppred": posterior predictions
\item "pq": posterior quantiles
\item "q_true": true quantiles
\item "r2_loo": LOO R-squared
\item "r2_newdata": R-squared on new data
\item "residuals": model residuals
\item "rhat": R-hat diagnostic
\item "rmse_loo": LOO root mean square error
\item "rmse_newdata": RMSE on new data
\item "rmse_s": scaled RMSE
\item "rstar": R* diagnostic
\item "time_sampling": sampling time
\item "time_total": total time
\item "time_warmup": warmup time
\item "y": observed values
}

Note,that not all identifiers are supported for each input class.
}
\examples{
\dontrun{
# Requires brms package and a fitted model
fit <- brms::brm(y ~ x, data = mydata)
metric_list_handler(fit, c("elpd_loo", "rhat"), data_gen_output, fit_conf)
}
}
