% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_list_handler.R
\name{metric_list_handler}
\alias{metric_list_handler}
\title{Title}
\usage{
metric_list_handler(
  fit,
  numeric_metrics,
  predictive_metrics,
  testing_data,
  ...
)
}
\arguments{
\item{fit}{}
}
\description{
Title
}
