% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_lookup.R
\name{metric_lookup}
\alias{metric_lookup}
\title{Access metrics with string identifiers}
\usage{
metric_lookup(
  metric,
  fit = NULL,
  draws = NULL,
  testing_data = NULL,
  variables = NULL,
  references = NULL,
  threshold = 0.7,
  psis_object = NULL,
  ppred = NULL,
  quantiles = NULL,
  data_gen_output = NULL,
  fit_conf = NULL,
  ...
)
}
\arguments{
\item{metric}{A string that identifies a supported metric}
}
\value{
The function corresponding to the identifier string.
}
\description{
This function is mostly a helper function that maps identifier strings to
metrics for convenient use in some contexts.
}
