% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{nodes_by_depth}
\alias{nodes_by_depth}
\title{Topological sort of response variables by dependency depth (Kahn's
algorithm). Nodes with no incoming edges (predictors that depend on no other
response) come first.}
\usage{
nodes_by_depth(adj_matrix)
}
\description{
Topological sort of response variables by dependency depth (Kahn's
algorithm). Nodes with no incoming edges (predictors that depend on no other
response) come first.
}
\keyword{internal}
