% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null-coalescing}
\alias{null-coalescing}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to check}

\item{y}{Default value if x is NULL}
}
\value{
\code{x} if not NULL, otherwise \code{y}
}
\description{
Returns \code{x} if not NULL, otherwise \code{y}.
}
\keyword{internal}
