% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-built-in.R
\name{PosteriorMeanMetric}
\alias{PosteriorMeanMetric}
\title{Posterior Mean Metric}
\usage{
PosteriorMeanMetric(
  name = character(0),
  needs = character(0),
  required = FALSE
)
}
\description{
Posterior mean estimates for parameters.
}
