% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.R
\name{qbetaprime}
\alias{qbetaprime}
\title{Quantile function of the Beta-Prime distributions using mean parametrization}
\usage{
qbetaprime(p, mu, phi)
}
\arguments{
\item{p}{Probabilities, for which to calculate the quantiles}

\item{mu}{Mean, mu > 0.}

\item{phi}{Precision, phi > 0.}
}
\value{
Quantiles of the beta-prime distribution, given p, mu and phi
}
\description{
Quantile function of the Beta-Prime distributions using mean parametrization
}
\examples{
x <- seq(from = 0, to = 1, length.out = 100)
plot(x, bayesim::qbetaprime(x, mu = 1, phi = 2), type = "l")
}
