% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\docType{data}
\name{RETAIN_OPTIONS}
\alias{RETAIN_OPTIONS}
\title{Valid retention options}
\format{
Character vector
}
\usage{
RETAIN_OPTIONS
}
\description{
Character vector of all valid retention option names that can be used
in retention specifications.
}
\keyword{internal}
