% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\docType{package}
\name{retention}
\alias{retention}
\title{Memory Management and Retention Policies}
\description{
Functions for controlling which artifacts are retained in simulation
results to manage memory usage during large-scale simulations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sims1253.github.io/bayesim}
  \item \url{https://github.com/sims1253/bayesim}
  \item Report bugs at \url{https://github.com/sims1253/bayesim/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Scholz \email{research.scholz@mailbox.org}

Authors:
\itemize{
  \item Maximilian Scholz \email{research.scholz@mailbox.org}
  \item Paul-Christian Bürkner \email{paul.buerkner@gmail.com}
}

}
\keyword{internal}
