% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\docType{data}
\name{RETENTION_PROFILES}
\alias{RETENTION_PROFILES}
\title{Retention profiles}
\format{
Named list of character vectors
}
\usage{
RETENTION_PROFILES
}
\description{
Pre-defined retention profiles for common use cases. Each profile is a
character vector of retention options.
}
\keyword{internal}
