% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_frechet.R
\name{rfrechet_custom}
\alias{rfrechet_custom}
\title{Frechet RNG function in Mean parametrisation}
\usage{
rfrechet_custom(n, mu, nu)
}
\arguments{
\item{mu}{Mean}

\item{nu}{Shape}
}
\value{
n samples as Frechet-distribution
}
\description{
Frechet RNG function in Mean parametrisation
}
