% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_handler.R
\name{rmse}
\alias{rmse}
\title{Calculate the root-mean-squared-error for given y and yrep.}
\usage{
rmse(y, yrep, weights = NULL)
}
\arguments{
\item{y}{Vector of observed values}

\item{yrep}{Matrix of predicted values (draws x observations)}

\item{weights}{PSIS weights (optional)}
}
\value{
rmse for the given y and yrep vectors
}
\description{
If psis-weights are supplied, the corresponding psis-rmse is returned.
}
\examples{
\dontrun{
y <- c(1, 2, 3)
yrep <- matrix(rnorm(300), nrow = 100, ncol = 3)
rmse(y, yrep)
}
}
