% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-built-in.R
\name{RmseMetric}
\alias{RmseMetric}
\title{RMSE Metric}
\usage{
RmseMetric(name = character(0), needs = character(0), required = FALSE)
}
\description{
Root Mean Square Error between predictions and true values.
}
