% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-built-in.R
\name{RmseMetric}
\alias{RmseMetric}
\alias{rmse_metric}
\title{RMSE Metric}
\usage{
RmseMetric(name = character(0), needs = character(0), required = FALSE)

rmse_metric(name = "rmse")
}
\arguments{
\item{name}{Character string naming the metric. Defaults to "rmse".}

\item{needs}{Character vector of required capabilities. Defaults to "predictions".}

\item{required}{Logical indicating if metric failure causes task failure. Defaults to FALSE.}
}
\value{
An RmseMetric object.
}
\description{
Root Mean Square Error between predictions and true values.

Constructor for RmseMetric.
}
\examples{
rmse_metric()
}
\keyword{internal}
