% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{rsimplex}
\alias{rsimplex}
\title{Simplex RNG function in Median parametrisations.}
\usage{
rsimplex(n, mu, sigma)
}
\arguments{
\item{n}{Number of samples to draw, as a natural number scalar.}

\item{mu}{Mean parameter, mu e (0, 1)}

\item{sigma}{shape parameter, Sigma unbound}
}
\value{
n samples in Simplex distribution.
}
\description{
Simplex RNG function in Median parametrisations.
}
\examples{
hist(rsimplex(10000, mu=0.7, sigma=2))
}
